+++ scraper/doctolib/doctolib_center_scrap.py
@@ -1,10 +1,6 @@
 import multiprocessing
-from time import sleep, time
-from scraper.pattern.scraper_result import (
-    DRUG_STORE,
-    GENERAL_PRACTITIONER,
-    VACCINATION_CENTER,
-)
+
+from scraper.pattern.scraper_result import VACCINATION_CENTER
 from utils.vmd_config import get_conf_platform, get_conf_inputs
 from utils.vmd_utils import departementUtils, format_phone_number
 from utils.vmd_logger import get_logger
@@ -16,7 +12,6 @@ import requests
 import json
 from urllib import parse
 import requests
-import os
 import re
 from unidecode import unidecode
