+++ scraper/keldoc/keldoc.py
@@ -1,6 +1,5 @@
 import os
 import logging
-from datetime import datetime, timedelta
 
 import httpx
