+++ scraper/mapharma/mapharma.py
@@ -7,15 +7,14 @@ import logging
 from datetime import date, datetime, timedelta
 from dateutil.parser import isoparse
 from pytz import timezone
-from urllib.parse import parse_qs
-from bs4 import BeautifulSoup
 from pathlib import Path
 from urllib import parse
 from typing import Optional
 
 from scraper.pattern.scraper_request import ScraperRequest
 from scraper.pattern.scraper_result import DRUG_STORE
-from scraper.pattern.center_info import get_vaccine_name, Vaccine, INTERVAL_SPLIT_DAYS, CHRONODOSES
+from scraper.pattern.center_info import INTERVAL_SPLIT_DAYS, CHRONODOSES
+from scraper.pattern.vaccine import get_vaccine_name
 from utils.vmd_config import get_conf_platform
 from utils.vmd_utils import departementUtils
 from scraper.profiler import Profiling
