+++ scraper/opendata/opendata.py
@@ -4,7 +4,7 @@ import json
 from typing import Iterator
 import requests
 
-from utils.vmd_logger import enable_logger_for_production, get_logger
+from utils.vmd_logger import get_logger
 from utils.vmd_utils import fix_scrap_urls
 
 logger = get_logger()
